backend/CryptoWalletService/requirements.txt
django
djangorestservice
django-angular
google-cloud-datastore
